+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   * Stop including query params when completing pageUrl macro
   * Hide the close button if the MiniReel is not skippable and it is
     running as an interstitial
+  * [DEV]: Allow a subset of tests to be run with ```grunt tdd``` via
+    the ```--only``` option.
 
 ## v0.21.0 (July 30, 2015)
 * *[v0.21.0-rc1]*